+++ packages/vue/src/index.ts
@@ -1,2 +1,4 @@
 import CuriPlugin from './plugin';
-export default CuriPlugin;
+import installCuri from './install';
+
+export { CuriPlugin, installCuri };