+++ packages/vue/src/plugin.ts
@@ -1,5 +1,5 @@
 import { CuriConfig } from '@curi/core';
-import { PluginObject, PluginFunction } from 'vue';
+import Vue, { PluginObject, PluginFunction, VueConstructor } from 'vue';
 
 import Link from './Link';
 import Block from './Block';
@@ -9,14 +9,28 @@ export interface CuriPluginOptions {
   curi: CuriPluginObject;
 }
 
+interface VueUtil {
+  defineReactive(
+    obj: Object,
+    key: string,
+    val: any,
+    customSetter?: Function,
+    shallow?: boolean
+  ): any;
+}
+
+interface VueWithUtil extends VueConstructor<Vue> {
+  util: VueUtil;
+}
+
 const CuriPlugin: PluginObject<CuriPluginOptions> = {
-  install: function(_Vue, options: CuriPluginOptions) {
+  install: function(_Vue: VueWithUtil, options: CuriPluginOptions) {
     _Vue.component(Link.name, Link);
     _Vue.component(Block.name, Block);
 
     _Vue.mixin({
-      created: function() {
-        this.$curi = options.curi;
+      beforeCreate: function() {
+        _Vue.util.defineReactive(this, '$curi', options.curi);
       }
     });
   }