+++ packages/vue/types/install.d.ts
@@ -1,0 +1,3 @@
+import Vue from 'vue';
+import { CuriConfig } from '@curi/core';
+export default function install(_Vue: typeof Vue, config: CuriConfig): void;